Extract test_invalid_configurations and test_valid_configurations to separate file
