+++ tests/rest_api/test_configurations.py
@@ -1,0 +1,14 @@
+import pytest
+
+from sources.rest_api import rest_api_source
+from .source_configs import VALID_CONFIGS, INVALID_CONFIGS
+
+@pytest.mark.parametrize("expected_message, exception, invalid_config", INVALID_CONFIGS)
+def test_invalid_configurations(expected_message, exception, invalid_config):
+    with pytest.raises(exception, match=expected_message):
+        rest_api_source(invalid_config)
+
+
+@pytest.mark.parametrize("valid_config", VALID_CONFIGS)
+def test_valid_configurations(valid_config):
+    rest_api_source(valid_config)
