+++ tests/rest_api/test_rest_api_source_offline.py
@@ -13,7 +13,6 @@ from sources.rest_api import (
 )
 from sources.rest_api.paginators import BasePaginator
 
-from .source_configs import VALID_CONFIGS, INVALID_CONFIGS
 
 
 def test_load_mock_api(mock_api_server):
@@ -289,14 +288,3 @@ def test_load_mock_api_typeddict_config(mock_api_server):
 
     assert table_counts["posts"] == 100
     assert table_counts["post_comments"] == 5000
-
-
-@pytest.mark.parametrize("expected_message, exception, invalid_config", INVALID_CONFIGS)
-def test_invalid_configurations(expected_message, exception, invalid_config):
-    with pytest.raises(exception, match=expected_message):
-        rest_api_source(invalid_config)
-
-
-@pytest.mark.parametrize("valid_config", VALID_CONFIGS)
-def test_valid_configurations(valid_config):
-    rest_api_source(valid_config)
